requirements.txt
torch
torchvision
numpy
matplotlib
pytorch-fid
Pillow